Changed ruby server to include Content-Type header with the value 'application/json'. This allowed the Javascript client to properly recognize that it was receiving JSON data back from the server.
I also added the initial files for the Javascript client. It's a
node project that contains a module for accessing the web service
(both the Javascript server and the ruby server) and a set of test
cases that uses the mocha testing framework.

To invoke the tests, you type "mocha" in the root level of the client
directory. In order for that to work, you must intall mocha by running
"npm install" in the root level of the client directory and then
making sure that the installed mocha executable is in your path.